+++ Src/CopernicaNET/Data/DatabaseCalls.cs
@@ -10,23 +10,23 @@ namespace Arlanet.CopernicaNET.Data
     {
         public static class POST
         {
-            public static string DeleteProfile = "profile/{0}?access_token={1}";
-            public static string CreateProfile = "database/{0}/profiles?access_token={1}";
-            public static string CreateSubProfile = "profile/{0}/subprofiles/{1}?access_token={2}";
-            public static string CreateDatabase = "databases?access_token={0}";
+            public const string DeleteProfile = "profile/{0}?access_token={1}";
+            public const string CreateProfile = "database/{0}/profiles?access_token={1}";
+            public const string CreateSubProfile = "profile/{0}/subprofiles/{1}?access_token={2}";
+            public const string CreateDatabase = "databases?access_token={0}";
         }
 
         public static class GET
         {
-            public static string GetProfileByField = "database/{0}/profiles?access_token={1}&fields[]={2}=={3}";
-            public static string GetProfileByFields = "database/{0}/profiles?access_token={1}&fields[]={2}";
-            public static string GetProfileFields = "database/{0}/fields?access_token={1}";
-            public static string GetSubProfileFields = "collection/{0}/fields?access_token={1}";
+            public const string GetProfileByField = "database/{0}/profiles?access_token={1}&fields[]={2}=={3}";
+            public const string GetProfileByFields = "database/{0}/profiles?access_token={1}&fields[]={2}";
+            public const string GetProfileFields = "database/{0}/fields?access_token={1}";
+            public const string GetSubProfileFields = "collection/{0}/fields?access_token={1}";
         }
 
         public static class PUT
         {
-            public static string UpdateProfile = "database/{0}/profiles?access_token={1}&fields[]={2}=={3}";
+            public const string UpdateProfile = "database/{0}/profiles?access_token={1}&fields[]={2}=={3}";
         }
     }
 }
